@@ -18,8 +18,9 @@
 package org.dmfs.webcal.utils.color;
 
 import android.content.Context;
-import android.support.annotation.ColorRes;
-import android.support.v4.content.ContextCompat;
+
+import androidx.annotation.ColorRes;
+import androidx.core.content.ContextCompat;
 
 
 /**
