@@ -19,9 +19,10 @@ package org.dmfs.webcal.utils.color;
 
 import android.content.Context;
 import android.content.res.Resources;
-import android.support.annotation.AttrRes;
-import android.support.annotation.ColorInt;
 import android.util.TypedValue;
+
+import androidx.annotation.AttrRes;
+import androidx.annotation.ColorInt;
 
 
 /**
